moved current contents into server folder
